Split unit tests from type_checker.rs for consistency
Now it's clear that main logic doesn't have any unwrap() or panic!().
