+++ parser/src/type_checker.rs
@@ -454,16 +454,4 @@ fn binary_cmp_type(lhs: &TypeDecl, rhs: &TypeDecl) -> Result<TypeDecl, ()> {
 }
 
 #[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn test_cast() {
-        let span = Span::new("var a: i64; a as i32");
-        let (_, ast) = crate::parser::source(span).unwrap();
-        assert_eq!(
-            type_check(&ast, &mut TypeCheckContext::new(None)).unwrap(),
-            TypeDecl::I32
-        );
-    }
-}
+mod test;
